tests fixed, refactoring
